script_test: Move helpers out
Make script_test smaller and more readable.
Move the terminal emulator helper out into a separate package,
and move the custom commands out into their own functions.

Also, compute the default environment in advance insteda of in setup.
